python_learning/gcp_python_projects/01_create_gcp_project_with_python/requirements.txt
google-cloud-resource-manager==1.7.0
google-auth==2.0.0
google-api-python-client==2.19.0
